refactor: enum 클래스인 Ranking 따로 분리
