+++ src/main/java/lotto/domain/constant/Ranking.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.constant;
 
 import java.util.List;
