+++ src/main/java/lotto/service/LottoService.java
@@ -4,10 +4,10 @@ import java.util.List;
 import lotto.domain.Lotto;
 import lotto.domain.Lottos;
 import lotto.domain.Profit;
-import lotto.domain.Ranking;
 import lotto.domain.Result;
 import lotto.domain.WinningLotto;
 import lotto.domain.checker.WinningChecker;
+import lotto.domain.constant.Ranking;
 import lotto.domain.extractor.LottoExtractor;
 import lotto.io.input.InputHandler;
 import lotto.io.output.PurchasePrintHandler;
